apps/market/pyproject.toml
[project]
name = "market"
version = "0.0.0"
description = "Project market. The online store"
readme = "README.md"
dependencies = [
    "punq>=0.7.0",
    "jinja2>=3.1.6",
    "uvicorn",
    "python-dotenv>=1.2.2",
    "starlette>=0.52.1",
    "uvicorn>=0.42.0",
    "grpc_buybikeshop_python",
]

[tool.uv.sources]
grpc_buybikeshop_python = { workspace = true }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
